incremental in moving to a common collection data store library
